onchain_control: replay to onchaind after we restart.
When onchaind sees a transaction removed from the blockchain, it is
killed.

We replay any notification once all the blocks are removed, which
restarts the onchaind; and as we restart from scratch, the channel
status is always returned to state FUNDING_SPEND_SEEN, so the FIXME
comment is not actually correct.

Fixes: #551
Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>