gossipd/lightningd: use wireaddr_or_sockname.
This replacement is a little menial, but it explicitly catches all
the places where we allow a local socket.

The detection of "valid address" is now more complex:

	p->addr.is_sockname || p->addr.u.wireaddr.type != ADDR_TYPE_PADDING

But most places we do this, we should audit: I'm pretty sure we can't
get an invalid address any more from gossipd (they may be in db, but
we should fix that too).

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>