gossipd: don't send gossip stream, let per-peer daemons read it themselves.
Keeping the uintmap ordering all the broadcastable messages is expensive:
130MB for the million-channels project.  But now we delete obsolete entries
from the store, we can have the per-peer daemons simply read that sequentially
and stream the gossip itself.

This is the most primitive version, where all gossip is streamed;
successive patches will bring back proper handling of timestamp filtering
and initial_routing_sync.

We add a gossip field to track what's happening with our gossip
streaming: currently it's always initialized in openingd, but once we
handle timestamps that will be deferred until the first filter is
sent.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>