gossipd/gossip_store: disable broken upgrade code.
It assumed that the channel_update immediately followed the channel_announcement,
but it doesn't in real gossip_stores, since it's probably been updated since then.

The complexity of all this isn't worthwhile, though it means we need to be more
selective on gossip after an upgrade to avoid a flood.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>