channeld: tell gossipd about new channel earlier.
We're about to change this to go via lightningd, not direct to gossipd,
and connectd will send gossip direct to gossipd.

Once we do this, it introduces a race.  The peer sends funding_locked
then immediately sends an update_channel.  channeld used to process
the funding_locked from the peer, tell gossipd about the new
channel, then finally forward the channel_update.

With the new architecture, we can have the channel_update hit gossipd
before we've told it about the channel.  It ignores the channel_update.

Better is to tell gossipd as soon as *we* are ready to send
funding_locked; this gives a round trip time at least, before gossipd
could get the update.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>