renepay/flow: use a hashtable for extra information.
Instead of an array of flow amounts and an array of known capacities,
we want to keep extra data on a handful of channels (including existing
flows and capacities, but it will be more).

Turns out it's best to push this down into flow.c, since it's common
information everyone will want.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>